Market price history

Download a data frame of contract prices over time, either hourly or daily. This chart data does not include a market name or numeric contract IDs, which can be obtained from the list of all open_markets().

market_history(mid, hourly = FALSE)

Arguments

mid
Market ID.

hourly
24 hourly rows per contract? If not, 90 daily rows.

Format

A tibble with 8 variables:

time

The hour or day of price

mid

Market ID

contract

Question answers

open

Price at start of hour/day

close

Price at end of hour/day

low

Low price of hour/day

high

High price of hour/day

volume

Number of shares traded per hour/day

Value

A data frame of market contract prices over time.

Examples

market_history(mid = 3698)

#> # A tibble: 2,127 x 8
#>    time                  mid contract    open  high   low close volume
#>    <dttm>              <int> <chr>      <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 2020-04-17 00:00:00  3698 Harris      0.02  0.02  0.01  0.02  13783
#>  2 2020-04-17 00:00:00  3698 Warren      0.01  0.01  0.01  0.01  16188
#>  3 2020-04-17 00:00:00  3698 Booker      0.01  0.01  0.01  0.01     10
#>  4 2020-04-17 00:00:00  3698 Kasich      0.01  0.01  0.01  0.01     50
#>  5 2020-04-17 00:00:00  3698 Biden       0.45  0.45  0.44  0.45  26009
#>  6 2020-04-17 00:00:00  3698 Sanders     0.01  0.01  0.01  0.01    798
#>  7 2020-04-17 00:00:00  3698 Pence       0.02  0.02  0.01  0.02   5092
#>  8 2020-04-17 00:00:00  3698 Trump       0.5   0.52  0.5   0.51  37443
#>  9 2020-04-17 00:00:00  3698 Gillibrand  0.01  0.01  0.01  0.01      0
#> 10 2020-04-17 00:00:00  3698 Cuomo       0.03  0.04  0.03  0.03   2069
#> # … with 2,117 more rows

market_history(mid = 6653, hourly = TRUE)

#> # A tibble: 384 x 8
#>    time                  mid contract         open  high   low close volume
#>    <dttm>              <int> <chr>           <dbl> <dbl> <dbl> <dbl>  <dbl>
#>  1 2020-07-15 18:00:00  6653 GOP by 0 - 9     0.05  0.06  0.05  0.05    186
#>  2 2020-07-15 18:00:00  6653 GOP by 10 - 29   0.05  0.06  0.05  0.06    200
#>  3 2020-07-15 18:00:00  6653 GOP by 30 -59    0.05  0.05  0.05  0.05    105
#>  4 2020-07-15 18:00:00  6653 Dems by 60 - 99  0.09  0.1   0.09  0.09    911
#>  5 2020-07-15 18:00:00  6653 GOP by 60 - 99   0.07  0.07  0.07  0.07    350
#>  6 2020-07-15 18:00:00  6653 GOP by 280+      0.03  0.03  0.02  0.03   1017
#>  7 2020-07-15 18:00:00  6653 Dems by 1 - 9    0.03  0.03  0.03  0.03    325
#>  8 2020-07-15 18:00:00  6653 Dems by 10 - 29  0.05  0.05  0.05  0.05    295
#>  9 2020-07-15 18:00:00  6653 Dems by 30 -59   0.06  0.07  0.06  0.06   1125
#> 10 2020-07-15 18:00:00  6653 Dems by 280+     0.17  0.17  0.16  0.16    527
#> # … with 374 more rows